requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.14
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.0
freetype-py==2.5.1
fsspec==2024.9.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.4
keras==2.13.1
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
pillow==10.4.0
protobuf==3.20.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyglet==2.0.17
pygltflib==1.16.2
PyOpenGL==3.1.0
pyparsing==3.1.4
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sounddevice==0.5.0
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
trimesh==4.4.9
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.100
ultralytics-thop==2.0.8
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
